Fix #12213: Incorrect viewport invalidation on load (#12232)
Mistake made during implementation but issue only surfaces recently when the types were increased to 32bit. Removed #12210 as this was hiding the source issue and is no longer required.

+++ src/openrct2/interface/Viewport.cpp
@@ -543,9 +543,8 @@ void viewport_update_position(rct_window* window)
 
     viewport_set_underground_flag(0, window, viewport);
 
-    // The midpoint relies on the overflow of int16_t to properly load a save on midscreen
-    auto viewportMidPoint = ScreenCoordsXY{ static_cast<int16_t>(window->savedViewPos.x + viewport->view_width / 2),
-                                            static_cast<int16_t>(window->savedViewPos.y + viewport->view_height / 2) };
+    auto viewportMidPoint = ScreenCoordsXY{ window->savedViewPos.x + viewport->view_width / 2,
+                                            window->savedViewPos.y + viewport->view_height / 2 };
 
     auto mapCoord = viewport_coord_to_map_coord(viewportMidPoint, 0);
